feat(modelos): Refactor LogisticaInputInserter to handle Demanda and Oferta inputs

See merge request apps/operaciones-industriales/blending-harina-2/blending-harina-2-backend!30

+++ Function.Blending.Opt/Application/Features/LogEjecucion/DTOs/Responses/Input/LogInpFilCapacidadDto.cs
@@ -2,6 +2,6 @@
 
 public sealed record LogInpFilCapacidadDto
 {
-  public int Cantidad { get; init; }
-  public int Capacidad { get; init; }
+  public decimal Cantidad { get; init; }
+  public decimal Capacidad { get; init; }
 }
